Rename AudioSink API and update CI runner
