+++ .github/workflows/ci.yml
@@ -10,7 +10,7 @@ on:
 jobs:
   build-macos:
     name: macOS (Swift 6.2)
-    runs-on: macos-15
+    runs-on: macos-26
     steps:
       - uses: actions/checkout@v4
